Xi/xkb: fix bit shifting by CWE-190 Integer Overflow or Wraparound
References:
- https://cwe.mitre.org/data/definitions/190.html

Example real CVE vulns with Integer Overflow:
- https://nvd.nist.gov/vuln/detail/CVE-2022-49748
- https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git/commit/?id=08245672cdc6505550d1a5020603b0a8d4a6dcc7

+++ Xi/xiproperty.c
@@ -272,7 +272,7 @@ get_property(ClientPtr client, DeviceIntPtr dev, Atom property, Atom type,
 
     /* Return type, format, value to client */
     n = (prop_value->format / 8) * prop_value->size;    /* size (bytes) of prop */
-    ind = offset << 2;
+    ind = (unsigned long)offset << 2;
 
     /* If offset is invalid such that it causes "len" to
        be negative, it's a value error. */
